fraud-radar/requirements.txt
absl-py==2.1.0
ace_tools==0.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.1
arabic-reshaper==3.0.0
argcomplete==3.5.3
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
branca==0.8.1
bs4==0.0.2
cachelib==0.9.0
cachetools==5.5.1
camelot-py==0.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
clickclick==20.10.2
cloudpickle==3.1.1
cmdstanpy==1.2.5
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.12
dacite==1.8.1
dask==2025.1.0
db-dtypes==1.4.1
debugpy==1.8.9
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==37.1.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastjsonschema==2.21.1
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
folium==0.19.4
fonttools==4.55.1
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud==0.34.0
google-cloud-bigquery==3.29.0
google-cloud-bigquery-storage==2.28.0
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
holidays==0.67
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.7.0
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
libclang==18.1.1
lightgbm==4.5.0
limits==4.0.1
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
lockfile==0.12.2
lxml==5.3.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
ml-dtypes==0.4.1
more-itertools==10.6.0
multidict==6.1.0
multimethod==1.12
multitasking==0.0.11
mysql-connector-python==9.2.0
namex==0.0.8
narwhals==1.29.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandas-profiling==3.6.6
parso==0.8.4
partd==1.4.2
passlib==1.7.4
pathspec==0.12.1
patsy==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.17.9
pendulum==3.0.0
phik==0.12.4
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
pmdarima==2.0.3
prison==0.2.1
prompt_toolkit==3.0.48
propcache==0.2.1
prophet==1.1.6
proto-plus==1.26.0
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-extra-types==2.10.2
pydantic-settings==2.8.0
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.11.2
PyMySQL==1.1.1
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyspark==3.5.4
python-bidi==0.6.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyWavelets==1.8.0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
setproctitle==1.3.4
shap==0.46.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
stanio==0.5.1
starlette==0.45.3
statsmodels==0.14.4
sweetviz==2.3.1
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
ucimlrepo==0.0.7
ujson==5.10.0
universal_pathlib==0.2.6
urllib3==2.2.3
uvicorn==0.34.0
visions==0.7.6
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0
Werkzeug==2.2.3
widgetsnbextension==4.0.13
wirerope==1.0.0
wordcloud==1.9.4
wrapt==1.17.0
WTForms==3.2.1
xgboost==2.1.4
xlrd==2.0.1
xyzservices==2025.1.0
yarl==1.18.3
ydata-profiling==4.12.0
yfinance==0.2.54
zipp==3.21.0
